Add game logic
Added new fucntion with comments that would simulate the game, human
vs computer and then output the score as well as what the computer chose
and whether that results in a win, loss or draw. Inserted a final function
call to run a single round of the game in each instance.